[1.4][sfSympalPlugin][1.0] Adding the idea of being able to include and render "objects". This extends the current idea of including "links" (relations to sfSympalContent rendered as links) and "assets" (relations to sfSympalAsset rendered however that asset type should be rendered).
This allows you to define any other type of thing to be replaced. Objects are selected in the same way that links and assets are selected and then are rendered via a partial.

I do have a few concerns about the location of some code - code without a home :) - and efficiency since we're always querying out for objects in these cases.


git-svn-id: http://svn.symfony-project.com/plugins/sfSympalPlugin/trunk@27331 ee427ae8-e902-0410-961c-c3ed070cd9f9