[1.4][sfSympalPlugin][1.0] Refactoring the text replacement system so that additional types beyond link and asset can be added. This creates a new plugin which handles this process.
The replacement system is also slot-agnostic, which means that it can be used as a standalone processor for text (though you'll lose efficiency compared to slots that use asset and link).

The pieces of the "object" replacement system have been removed and will be used for a standalone plugin that allows you to render any object via the [myObject:5] syntax.


git-svn-id: http://svn.symfony-project.com/plugins/sfSympalPlugin/trunk@27343 ee427ae8-e902-0410-961c-c3ed070cd9f9